Added placeId property to LocationInfo, setting the placeId field in the TopPlacesViewController, and the ability to fetch photos from a place based on placeId.

+++ Exercise5/Flickr/FlickrInformation.h
@@ -38,7 +38,7 @@ typedef enum {
 
 + (NSURL *)URLforTopPlaces;
 
-+ (NSURL *)URLforPhotosInPlace:(id)flickrPlaceId maxResults:(int)maxResults;
++ (NSURL *)URLforPhotosInPlace:(id)flickrPlaceId maxResults:(NSUInteger)maxResults;
 
 + (NSURL *)URLforPhoto:(NSDictionary *)photo format:(FlickrPhotoFormat)format;
